# resiresi-backend/openapi.yaml
openapi: "3.1.0"
info:
  title: ResiResi Restaurant Reservation API
  version: "1.0.0"
  description: |
    Backend reservation API for the ResiResi platform. Used as the APIblaze proxy reference origin.
    Authenticate with either `Authorization: Bearer <key>` or `x-api-key: <key>`.
servers:
  - url: https://backend.resiresi.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              example: not_found
            message:
              type: string
              example: Restaurant not found.
            request_id:
              type: string
              format: uuid
            details:
              type: object
              description: Zod flatten() output on validation errors
...
    ReservationStatus:
      type: string
      enum: [pending, confirmed, seated, completed, cancelled, no_show]
    Reservation:
      type: object
      required: [id, restaurant_id, diner_name, party_size, starts_at, ends_at, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        restaurant_id:
          type: string
          format: uuid
        table_id:
          type: string
          format: uuid
          nullable: true
        diner_user_id:
          type: string
          format: uuid
          nullable: true
        diner_name:
          type: string
        diner_email:
          type: string
          format: email
          nullable: true
        diner_phone:
          type: string
          nullable: true
        party_size:
          type: integer
          minimum: 1
          maximum: 50
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/ReservationStatus'
        notes:
          type: string
          nullable: true
        idempotency_key:
          type: string
          nullable: true
        created_by_key_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PageInfo:
      type: object
      properties:
        next_cursor:
          type: string
          nullable: true
          description: base64url cursor to pass as ?cursor= for the next page
        limit:
          type: integer
...
    ReservationList:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Reservation'
        page:
          $ref: '#/components/schemas/PageInfo'
...
paths:
...
  /restaurants/{restaurantId}/reservations:
    parameters:
      - name: restaurantId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      operationId: createReservation
      summary: Create reservation
      tags: [Reservations]
      parameters:
        - name: Idempotency-Key
          in: header
          schema:
            type: string
          description: Idempotency key — safe to retry with same key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [diner_name, party_size, starts_at]
              properties:
                table_id:
                  type: string
                  format: uuid
                diner_user_id:
                  type: string
                  format: uuid
                diner_name:
                  type: string
                diner_email:
                  type: string
                  format: email
                diner_phone:
                  type: string
                party_size:
                  type: integer
                  minimum: 1
                  maximum: 50
                starts_at:
                  type: string
                  format: date-time
                ends_at:
                  type: string
                  format: date-time
                duration_minutes:
                  type: integer
                  minimum: 1
                  default: 90
                status:
                  $ref: '#/components/schemas/ReservationStatus'
                notes:
                  type: string
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        "200":
          description: Idempotency replay — reservation already exists
          headers:
            Idempotent-Replay:
              schema:
                type: string
                enum: ["true"]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        "400":
          description: Validation or semantic error (table_too_small, invalid_times)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Table conflict (double-booking)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      operationId: listReservations
      summary: List reservations
      tags: [Reservations]
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: status
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ReservationStatus'
          style: form
          explode: true
        - name: diner_email
          in: query
          schema:
            type: string
        - name: table_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: cursor
          in: query
          schema:
            type: string
          description: base64url cursor from previous response
      responses:
        "200":
          description: Reservation list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationList'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restaurants/{restaurantId}/reservations/{reservationId}:
    parameters:
      - name: restaurantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: reservationId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: getReservation
      summary: Get reservation
      tags: [Reservations]
      responses:
        "200":
          description: Reservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: updateReservation
      summary: Update reservation
      tags: [Reservations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                table_id:
                  type: string
                  format: uuid
                  nullable: true
                diner_name:
                  type: string
                diner_email:
                  type: string
                  format: email
                diner_phone:
                  type: string
                party_size:
                  type: integer
                  minimum: 1
                  maximum: 50
                starts_at:
                  type: string
                  format: date-time
                ends_at:
                  type: string
                  format: date-time
                status:
                  $ref: '#/components/schemas/ReservationStatus'
                notes:
                  type: string
      responses:
        "200":
          description: Updated reservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Table conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: cancelReservation
      summary: Cancel reservation (soft delete)
      tags: [Reservations]
      responses:
        "204":
          description: Cancelled
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...